+++ src/main/java/org/bytedeco/javacpp/tools/Token.java
@@ -48,10 +48,11 @@ class Token implements Comparable<Token> {
             SYMBOL     = 6;
 
     static final Token
-            EOF       = new Token(),
+            EOF       = new Token(-1, "EOF"),
             AUTO      = new Token(IDENTIFIER, "auto"),
             CONST     = new Token(IDENTIFIER, "const"),
             CONSTEXPR = new Token(IDENTIFIER, "constexpr"),
+            DECLTYPE  = new Token(IDENTIFIER, "decltype"),
             DEFAULT   = new Token(IDENTIFIER, "default"),
             DEFINE    = new Token(IDENTIFIER, "define"),
             IF        = new Token(IDENTIFIER, "if"),